+++ SDCTA/src/components/modal/styleObjects/modals.tsx
@@ -1,0 +1,47 @@
+export const exitButtonStyle = {
+    //location styling
+    position: 'absolute' as 'absolute',
+    left: '92.96%',
+    right: '3.2%',
+    top: '5.97%',
+    bottom: '88.36%',
+    color: 'black',
+};
+
+export const titleStyle = {
+    //location styling
+    width: "456px",
+    height: "49px",
+    lineHeight: '49px',
+    textAlign: "center",
+    order: 0,
+    zIndex: 0,
+    //font styling
+    fontFamily: 'sans-serif',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '39px',
+    color: "#C3272E",
+};
+
+export const buttonStyle = {
+    //location styling
+    width: '465px',
+    height: '48px',
+    mt: "40px",
+    order: 2,
+    //button styling
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: '0px',
+    gap: '10px',
+    background: '#C3272E',
+    //font styling
+    fontFamily: 'sans-serif',
+    fontSize: '18px',
+    textTransform: 'none',
+    //border styling
+    borderRadius: '4px',
+};
